Improved user experience, reduced actions amount
