+++ Apps.Okapi/Models/Requests/UploadFilesRequest.cs
@@ -1,0 +1,8 @@
+﻿using Blackbird.Applications.Sdk.Common.Files;
+
+namespace Apps.Okapi.Models.Requests;
+
+public class UploadFilesRequest
+{
+    public IEnumerable<FileReference> Files { get; set; }
+}
