+++ Apps.Okapi/Models/Responses/ExecuteProjectResponse.cs
@@ -1,0 +1,10 @@
+﻿using Blackbird.Applications.Sdk.Common;
+using Blackbird.Applications.Sdk.Common.Files;
+
+namespace Apps.Okapi.Models.Responses;
+
+public class ExecuteProjectResponse
+{
+    [Display("Output files")]
+    public IEnumerable<FileReference> OutputFiles { get; set; }
+}
